💄 (ui): Added Component for User Stats
